+++ database/migrations/2021_06_27_180320_create_userables_table.php
@@ -15,7 +15,7 @@ class CreateUserablesTable extends Migration
     {
         Schema::create('userables', function (Blueprint $table) {
             $table->unsignedBigInteger('user_id');
-            $table->unsignedInteger('userable_id');
+            $table->unsignedBigInteger('userable_id');
             $table->string('userable_type');
             $table->primary(['userable_id', 'userable_type']);
             $table->foreign('user_id', 'fk_user_userable')->references('id')->on('users');
